Throw IOException to prevent crash in LoginTask
Crash was caused in LoginTask when signing in with mode "email_only" when connected to wifi but no internet connectivity - the exact details of why this is then throwing an IOException but the same login with mode "email_password" does not will stay with the implementors of URLConnection.
The code is now just throwing the IOException to the next in the chain so they can handle it correctly.