Make request counter safer
Before, the request count would be raised instantly even if it was not entirely clear, whether a request would even be created. This could happen if e.g. the persistence was not available, or the persistence - for whatever reason - did not return anything to send.
Also the request counter was not decremented correctly if an exception occurred while sending.
This moves incrementing the request counter right to the point before sending starts, and implements a failsafe to decrement the request counter if an exception during sending prevents the reach of "onResponse()".