Remove `android:supportsRtl` attribute from AndroidManifest
The resources in the SDK might support RTL layouts, but the consuming app might not. In fact, when they explicitly set `supportsRtl` to `false` on their application, this causes a conflict when creating the merged manifest, causing the build to fail. So it is probably best not to set the attribute and ignore the lint warning.

Fixes #334.